requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2025.8.3
click==8.2.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.115.0
fastapi-utils==0.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
limits==5.5.0
openpyxl==3.1.5
packaging==25.0
passlib==1.7.4
psutil==5.9.8
psycopg==3.2.1
psycopg-binary==3.2.1
pyasn1==0.6.1
pydantic==2.8.2
pydantic-settings==2.10.1
pydantic_core==2.20.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
